Hook on schema editor changes when changing Meta Schema and fix typo
When loading a new Meta Schema, the schema editor is re-created and we forgot to hook again on thechanges to update the preview editor in real time. Also, destroyEditor method has been renamed destroy, but there still was a destroyEditor call, it has been correctly replaced with a call to destroy method.